feat: switch to near-runner-ava
uses in-progress version being built at https://github.com/near/runner-js/pull/64
