+++ examples/non-fungible-token/__tests__/nft_transfer_call.ava.ts
@@ -1,0 +1,118 @@
+import path from 'path';
+import { Gas } from 'near-units';
+import type { Token } from './utils';
+import { createRunner, TOKEN_ID } from './utils';
+
+const runner = createRunner(async ({ root, nft }) => ({
+  tokenReceiver: await root.createAndDeploy(
+    'token-receiver',
+    path.join(__dirname, '..', 'res', 'token_receiver.wasm'),
+    {
+      method: 'new',
+      args: { non_fungible_token_account_id: nft }
+    }
+  )
+}));
+
+runner.test('refund (no cross-contract call from receiver)', async (t, { root, nft, tokenReceiver }) => {
+  const tx = await root.call_raw(nft, 'nft_transfer_call', {
+    receiver_id: tokenReceiver,
+    token_id: TOKEN_ID,
+    memo: 'transfer and call',
+    msg: 'return-it-now',
+  }, {
+    attachedDeposit: '1',
+    gas: Gas.parse('31 Tgas'),
+  });
+
+  // Make sure all cross-contract calls worked as expected!
+  // This is needed because transfer_call gracefully ignores cross-contract call failures,
+  // so the transaction can pass even if cross-contract calls fail.
+  t.is(tx.errors.length, 0);
+
+  const { owner_id } =
+    await nft.view('nft_token', { token_id: TOKEN_ID }) as Token;
+  t.is(owner_id, root.accountId);
+});
+
+runner.test('refund after receiver makes cross-contract call', async (t, { root, nft, tokenReceiver }) => {
+  const tx = await root.call_raw(nft, 'nft_transfer_call', {
+    receiver_id: tokenReceiver,
+    token_id: TOKEN_ID,
+    memo: 'transfer and call',
+    msg: 'return-it-later',
+  }, {
+    attachedDeposit: '1',
+    gas: Gas.parse('41 Tgas'),
+  });
+
+  // Make sure all cross-contract calls worked as expected!
+  // This is needed because transfer_call gracefully ignores cross-contract call failures,
+  // so the transaction can pass even if cross-contract calls fail.
+  t.is(tx.errors.length, 0);
+
+  const { owner_id } =
+    await nft.view('nft_token', { token_id: TOKEN_ID }) as Token;
+  t.is(owner_id, root.accountId);
+});
+
+runner.test('success (no cross-contract call from receiver)', async (t, { root, nft, tokenReceiver }) => {
+  const tx = await root.call_raw(nft, 'nft_transfer_call', {
+    receiver_id: tokenReceiver,
+    token_id: TOKEN_ID,
+    memo: 'transfer and call',
+    msg: 'keep-it-now',
+  }, {
+    attachedDeposit: '1',
+    gas: Gas.parse('31 Tgas'),
+  });
+
+  // Make sure all cross-contract calls worked as expected!
+  // This is needed because transfer_call gracefully ignores cross-contract call failures,
+  // so the transaction can pass even if cross-contract calls fail.
+  t.is(tx.errors.length, 0);
+
+  const { owner_id } =
+    await nft.view('nft_token', { token_id: TOKEN_ID }) as Token;
+  t.is(owner_id, tokenReceiver.accountId);
+});
+
+runner.test('success after receiver makes cross-contract call', async (t, { root, nft, tokenReceiver }) => {
+  const tx = await root.call_raw(nft, 'nft_transfer_call', {
+    receiver_id: tokenReceiver,
+    token_id: TOKEN_ID,
+    memo: 'transfer and call',
+    msg: 'keep-it-later',
+  }, {
+    attachedDeposit: '1',
+    gas: Gas.parse('41 Tgas'),
+  });
+
+  // Make sure all cross-contract calls worked as expected!
+  // This is needed because transfer_call gracefully ignores cross-contract call failures,
+  // so the transaction can pass even if cross-contract calls fail.
+  t.is(tx.errors.length, 0);
+
+  const { owner_id } =
+    await nft.view('nft_token', { token_id: TOKEN_ID }) as Token;
+  t.is(owner_id, tokenReceiver.accountId);
+});
+
+runner.test('refund if receiver panics', async (t, { root, nft, tokenReceiver }) => {
+  const tx = await root.call_raw(nft, 'nft_transfer_call', {
+    receiver_id: tokenReceiver,
+    token_id: TOKEN_ID,
+    memo: 'transfer and call',
+    msg: 'keep-it-later',
+  }, {
+    attachedDeposit: '1',
+    gas: Gas.parse('40 Tgas'), // not enough gas!
+  });
+
+  t.is(tx.promiseErrors.length, 1);
+  t.regex(tx.promiseErrorMessages[0], /FunctionCallZeroAttachedGas/);
+
+  const { owner_id } =
+    await nft.view('nft_token', { token_id: TOKEN_ID }) as Token;
+  t.is(owner_id, root.accountId);
+});
